refactor: update import paths in Clock and OptionControls; remove unused useAudio and useTime hooks, and clean up utils and effects files
