+++ .gitignore
@@ -42,3 +42,4 @@ next-env.d.ts
 
 # Notes
 REFACTOR_PLAN.md
+architecture-review.md
