+++ src/app/Clock.tsx
@@ -3,8 +3,8 @@ import OptionControls from "./components/OptionControls";
 import AudioControls from "./components/AudioControls";
 import TimeDisplay from "./components/TimeDisplay";
 import AppHeader from "./components/AppHeader";
-import { useTime } from "./hooks/useTime";
-import { useAudio } from "./hooks/useAudio";
+import { useTime } from "../hooks/useTime";
+import { useAudio } from "../hooks/useAudio";
 
 const Clock = () => {
   const { time, mounted, selectedTimezone, setSelectedTimezone } = useTime();
