+++ src/hooks/useTime.ts
@@ -69,4 +69,4 @@ export const useTime = (): UseTimeReturn => {
     selectedTimezone,
     setSelectedTimezone,
   };
-};
+};
